refactor(outputs): close output boundary checkpoint
Why:
- keep output-format and review behavior inside output adapters instead of application services
- remove remaining output-specific naming from generic workflow and round scaffolding surfaces

What:
- add typed output workflow contracts and move CoinTracking baseline, screening, and reconcile logic into a package-style output adapter
- generalize render, baseline, staging, CLI, and round-log surfaces around output adapters
- update roadmap to leave only the remaining source-intake boundary work recorded

Checks:
- uv run pre-commit run markdownlint --all-files
- uv run python -m tools.run_quality_gates --full-tests
