+++ src/crypto_reconciliation/adapters/outputs/cointracking_csv/__init__.py
@@ -1,0 +1,5 @@
+"""CoinTracking CSV output adapter package."""
+
+from .schema import COINTRACKING_HEADER
+
+__all__ = ["COINTRACKING_HEADER"]
