+++ src/crypto_reconciliation/application/services/rounds.py
@@ -14,7 +14,7 @@ ROUND_LOG_HEADER = (
     "source",
     "date",
     "goal",
-    "cointracking_change",
+    "output_change",
     "exports_captured",
     "issues_opened_or_closed",
     "gate_result",
@@ -109,7 +109,7 @@ def create_round_log_entry(
         "source": request.source,
         "date": today.isoformat(),
         "goal": goal,
-        "cointracking_change": "",
+        "output_change": "",
         "exports_captured": str(verification_dir.relative_to(request.workspace_root)),
         "issues_opened_or_closed": "",
         "gate_result": "pending",
