+++ src/crypto_reconciliation/ports/__init__.py
@@ -10,18 +10,22 @@ from .adapters import (
 )
 from .ai import ModelGateway, ReviewRequest, ReviewResponse
 from .artifacts import ArtifactStorePort
+from .output_workflows import BaselineArtifacts, OverlapResult, ScreeningResult
 from .storage import StoragePort
 from .workspace import WorkspaceLocator, WorkspaceRepository
 
 __all__ = [
     "ArtifactStorePort",
+    "BaselineArtifacts",
     "ModelGateway",
     "NormalizationResult",
     "OutputAdapter",
     "OutputAdapterRegistryPort",
+    "OverlapResult",
     "RenderedArtifact",
     "ReviewRequest",
     "ReviewResponse",
+    "ScreeningResult",
     "SourceAdapter",
     "SourceAdapterRegistryPort",
     "StoragePort",
