+++ src/crypto_reconciliation/ports/output_workflows.py
@@ -1,11 +1,17 @@
-"""Staging workflow models."""
+"""Typed workflow artifacts owned by output adapters."""
 
 from __future__ import annotations
 
 from dataclasses import dataclass
 
-from crypto_reconciliation.application.services.overlap import OverlapResult
 from crypto_reconciliation.domain.models import IssueRecord
+from crypto_reconciliation.domain.types import JsonValue
+
+
+@dataclass(frozen=True)
+class OverlapResult:
+    summary: dict[str, JsonValue]
+    flagged_rows: tuple[dict[str, str], ...]
 
 
 @dataclass(frozen=True)
@@ -33,3 +39,14 @@ class ScreeningResult:
         if self.overlap_result is not None and self.overlap_result.summary["rows_flagged"]:
             reasons.append("overlap_review_required")
         return tuple(reasons)
+
+
+@dataclass(frozen=True)
+class BaselineArtifacts:
+    asset_snapshot_rows: list[dict[str, str]]
+    reconciliation_rows: list[dict[str, str]]
+    negative_balances: list[dict[str, str]]
+    source_activity_rows: list[dict[str, str]]
+    cad_flow_rows: list[dict[str, str]]
+    cad_balance_by_exchange_rows: list[dict[str, str]]
+    summary: dict[str, JsonValue]
