+++ tests/unit/application/services/test_rounds_service.py
@@ -84,7 +84,7 @@ def test_round_scaffolding_service_creates_round_dir_readme_and_log_entry(tmp_pa
             "source": "shakepay",
             "date": "2026-03-22",
             "goal": "Capture fresh verification exports after baseline repair",
-            "cointracking_change": "",
+            "output_change": "",
             "exports_captured": "working/verification/round_01",
             "issues_opened_or_closed": "",
             "gate_result": "pending",
@@ -128,7 +128,7 @@ def test_round_scaffolding_service_preserves_existing_round_log_rows(tmp_path: P
             "source",
             "date",
             "goal",
-            "cointracking_change",
+            "output_change",
             "exports_captured",
             "issues_opened_or_closed",
             "gate_result",
@@ -141,7 +141,7 @@ def test_round_scaffolding_service_preserves_existing_round_log_rows(tmp_path: P
                 "source": "cointracking",
                 "date": "2026-03-21",
                 "goal": "Existing goal",
-                "cointracking_change": "",
+                "output_change": "",
                 "exports_captured": "working/verification/round_00",
                 "issues_opened_or_closed": "",
                 "gate_result": "pass",
